+++ CHANGELOG.md
@@ -116,6 +116,7 @@
 - Dev: Decoupled reply parsing from `MessageBuilder`. (#5660, #5668)
 - Dev: Refactored IRC message building. (#5663)
 - Dev: Fixed some compiler warnings. (#5672)
+- Dev: Unified parsing of historic and live IRC messages. (#5678)
 
 ## 2.5.1